refactor(core): split edge_concurrent cascade helpers into sibling module
Move the five node-cascade cleanup helpers (collect_node_edges,
gather_affected_shards, cleanup_shard_edges, cleanup_cross_shard_edges,
deregister_edge_ids) out of edge_concurrent/mod.rs into a new cascade.rs
sibling as a split impl block. remove_node_edges and its lock acquisition
(edge_ids first, then shards ascending) stay in mod.rs, so lock ordering is
unchanged. Pure module reorganization: drops mod.rs file complexity from 54
to ~38 with no behaviour change. All 55 edge_concurrent tests pass.
